+++ db/migrate/20210216000757_create_tasks.rb
@@ -4,8 +4,8 @@ class CreateTasks < ActiveRecord::Migration[6.1]
       t.string :name
       t.references :calendar, index: true, foreign_key: {to_table: :calendars}
       t.string :desc
-      t.integer :start
-      t.integer :end
+      t.float :start
+      t.float :end
       t.integer :day
 
       t.timestamps
